requirements.txt
django
django-cors-headers
djangorestframework
tflite-runtime
numpy
gunicorn
pillow
whitenoise
